code adjustments
